Moved backbone and underscore to peerDependencies
A subtle bug occurs when you install two versions of backbone. One that
gets installed as a dependency to this package, and another that you
install to use in your project. (This is what happens with the default install
with browserify.)

When we filter and reset the filtered collection, Backbone uses
something like the following code:

    if (model instanceof Backbone.Model)

except that if you created the model with one version of backbone, and
the filtered collection was created referencing another, this will fail
since it's actually a different function in the prototype chain, even if
they're functionally identical.